Fixups to emf-config cli (#141)
When testing the emf-config cli I found a few things that need to be
fixed in order for deployment to work.

+++ emf-manager-cli/src/postgres.rs
@@ -47,7 +47,7 @@ pub enum Command {
 fn check_space(path: &str) -> Result<(), EmfManagerCliError> {
     let statvfs = statvfs(path)?;
 
-    let free = statvfs.blocks_free() * statvfs.fragment_size() / (1024 * 1024 * 1024);
+    let free = statvfs.blocks_free() as u64 * statvfs.fragment_size() / (1024 * 1024 * 1024);
 
     if free >= CHECK_SIZE_GB {
         Ok(())
